+++ utgb-core/src/main/java/org/utgenome/gwt/utgb/client/track/lib/BEDCanvasTrack.java
@@ -108,8 +108,8 @@ public class BEDCanvasTrack extends TrackBase {
 	public void draw() {
 		TrackWindow w = getTrackGroup().getTrackWindow();
 
-		long s = w.getStartOnGenome();
-		long e = w.getEndOnGenome();
+		int s = w.getStartOnGenome();
+		int e = w.getEndOnGenome();
 		int width = w.getWindowWidth() - 100;
 
 		geneCanvas.clear();
@@ -166,8 +166,8 @@ public class BEDCanvasTrack extends TrackBase {
 
 	public void update(TrackWindow newWindow) {
 		// retrieve gene data from the API
-		long s = newWindow.getStartOnGenome();
-		long e = newWindow.getEndOnGenome();
+		int s = newWindow.getStartOnGenome();
+		int e = newWindow.getEndOnGenome();
 		TrackGroupProperty prop = getTrackGroup().getPropertyReader();
 		String species = prop.getProperty(UTGBProperty.SPECIES);
 		String revision = prop.getProperty(UTGBProperty.REVISION);
